+++ bundles/net.tourbook.ui.tests/src/test/views/WorkbenchTests.java
@@ -13,7 +13,7 @@
  * this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110, USA
  *******************************************************************************/
-package basic;
+package views;
 
 import static org.junit.jupiter.api.Assertions.assertNotNull;
